[♻️ refactor/#107]: AuthService로 SignUpFilterService의 기능 통합
- `SignUpFilterService` 클래스 삭제 및 관련 기능 `AuthService`로 통합
- 사용자 필터 생성 및 연결 로직을 `AuthService`로 이전하여 서비스 계층 단순화
- 불필요한 메서드 및 인터페이스 제거로 코드 간소화
- `AuthController`에서 `SignUpFilterService` 의존성을 `AuthService`로 대체
