+++ src/main/java/org/terning/terningserver/repository/user/UserRepository.java
@@ -8,8 +8,6 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    Optional<User> findByAuthTypeAndRefreshToken(AuthType authType, String refreshToken);
-
     Optional<User> findByRefreshToken(String refreshToken);
 
     Optional<User> findByAuthId(String authId);
