+++ src/main/java/org/terning/terningserver/service/AuthService.java
@@ -1,7 +1,9 @@
 package org.terning.terningserver.service;
 
+import org.terning.terningserver.domain.Filter;
 import org.terning.terningserver.domain.enums.AuthType;
 import org.terning.terningserver.dto.auth.request.SignInRequestDto;
+import org.terning.terningserver.dto.auth.request.SignUpFilterRequestDto;
 import org.terning.terningserver.dto.auth.request.SignUpRequestDto;
 import org.terning.terningserver.dto.auth.request.SignUpWithAuthIdRequestDto;
 import org.terning.terningserver.dto.auth.response.AccessTokenGetResponseDto;
@@ -19,4 +21,9 @@ public interface AuthService {
     void withdraw(long userId);
 
     AccessTokenGetResponseDto reissueToken(String refreshToken);
+
+    Filter createAndSaveFilter(SignUpFilterRequestDto request);
+
+    void connectFilterToUser(Long userId, Long filterId);
+
 }
